+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Change log
 v0.1.13 <2018-xx-xx>
 --------------------
 
+- Add All zero synthesis filter.
 - Add ``levdur``.
 - Add tranposed synthesis filters (``mlsadft``, ``poledft``, ``mglsadft``, ``glsadft``)
 - Add missing high level synthesis filter class ``GLSADF``.